.cosmetics
git-svn-id: svn://ultimatepp.org/upp/trunk@10018 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Draw/Image.cpp
@@ -246,7 +246,7 @@ void Image::Serialize(Stream& s)
 			}
 			else {
 				Huge h;
-				if(!s.GetAll(h, len)) {
+				if(!s.GetAll(h, (size_t)len)) {
 					s.LoadError();
 					Clear();
 					return;
